move fetching rules within sources hook
